+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - For our new users we completely reworked the on-boarding experience.
  - We've made it a bit clearer in the app when you have unread notifications.
+ - The room creation flow has been improved for making private rooms.
 
 ## [1.9.0] - 2020-12-28